Flowsource-templates/PDLC-Agents/azure/Container-app/skeleton/containerapp/PDLC-Azure/requirements.txt
ag2==0.9.9
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
asyncpg==0.30.0
attrs==25.3.0
autogen==0.9.9
autogen-agentchat==0.7.4
autogen-core==0.7.4
autogen-ext==0.7.4
azure-ai-agents==1.0.0
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b11
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
fastapi==0.116.1
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
mcp==1.13.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
numpy==2.2.6
openai==1.79.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.2.3
pillow==12.1.1
propcache==0.3.1
protobuf==7.34.0rc2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.22
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.50.0
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.34.2
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
